Adicionando enum de "ponto de venda" em objetos de relatório
