+++ backend/src/main/java/br/com/chadschoperia/repository/reports/RevenueExpenseReportRepository.java
@@ -13,7 +13,7 @@ import java.util.List;
 @Repository
 public interface RevenueExpenseReportRepository extends JpaRepository<RevenueExpense, Long> {
 	@Query("SELECT new br.com.chadschoperia.service.reports.dto.RevenueExpenseReportDto" +
-			"(re.type, re.description, re.dateTime, re.value)" +
+			"(re.type, re.description, re.dateTime, re.value, re.sellingPoint)" +
 			" FROM RevenueExpense re " +
 			" WHERE (CAST(CAST(:#{#filter.minDate} AS char) AS timestamp) IS NULL OR CAST(:#{#filter.minDate} AS timestamp) <= re.dateTime) " +
 			" AND (CAST(CAST(:#{#filter.maxDate} AS char) AS timestamp) IS NULL OR CAST(:#{#filter.maxDate} AS timestamp) >= re.dateTime) " +
