+++ backend/src/main/java/br/com/chadschoperia/service/reports/dto/ClientExpesesReportDto.java
@@ -2,10 +2,10 @@ package br.com.chadschoperia.service.reports.dto;
 
 import br.com.chadschoperia.domain.entities.Client;
 import br.com.chadschoperia.domain.entities.ClientCard;
+import br.com.chadschoperia.domain.entities.ClientCardExpense;
 import br.com.chadschoperia.domain.enums.ClientCardStatusEnum;
+import br.com.chadschoperia.domain.enums.SellingPointEnum;
 import lombok.Getter;
-
-import java.time.LocalDateTime;
 
 @Getter
 public class ClientExpesesReportDto extends BaseReportDto {
@@ -16,9 +16,10 @@ public class ClientExpesesReportDto extends BaseReportDto {
 	private final Long clientCardId;
 	private final ClientCardStatusEnum clientCardStatus;
 	private final String clientCardrfid;
+	private final SellingPointEnum sellingPoint;
 
-	public ClientExpesesReportDto(String description, Double value, LocalDateTime dateTime, Client client, ClientCard card) {
-		super(description, value, dateTime);
+	public ClientExpesesReportDto(Client client, ClientCard card, ClientCardExpense expense) {
+		super(expense.getDescription(), expense.getValue(), expense.getDateTime());
 		this.clientId = client.getId();
 		this.clientName = client.getName();
 		this.clientTelephone = client.getTelephone();
@@ -26,5 +27,6 @@ public class ClientExpesesReportDto extends BaseReportDto {
 		this.clientCardId = card.getId();
 		this.clientCardStatus = card.getStatus();
 		this.clientCardrfid = card.getRfid();
+		this.sellingPoint = expense.getSellingPoint();
 	}
 }
