--- frontend/src/app/modules/report/modules/report-revenue/models/revenueExpenseReport.model.ts
@@ -1,12 +1,0 @@
-import { BaseReport } from '../../../models/base.report';
-
-export class RevenueExpenseReportModel extends BaseReport {
-	constructor(
-		description: string,
-		value: number,
-		dateTime: Date,
-		public type: 'REVENUE' | 'EXPENSE'
-	) {
-		super(description, value, dateTime);
-	}
-}
